Improve design of the calculation refresh banner
