+++ fdm-app/app/components/blocks/calculation/calculation-refresh-spinner.tsx
