+++ fdm-app/app/components/blocks/nutrient-advice/overview-columns.tsx
@@ -1,8 +1,8 @@
 import type { Column, ColumnDef, RowData } from "@tanstack/react-table"
 import { ArrowDown, ArrowUp } from "lucide-react"
 import { NavLink } from "react-router"
+import { CalculationRefreshSpinner } from "~/components/blocks/calculation/calculation-refresh-spinner"
 import { CultivationSuggestionBadge } from "~/components/blocks/cultivation/suggestion"
-import { CalculationRefreshSpinner } from "~/components/blocks/calculation-refresh-spinner"
 import { getCultivationColor } from "~/components/custom/cultivation-colors"
 import { Badge } from "~/components/ui/badge"
 import { Tooltip, TooltipContent, TooltipTrigger } from "~/components/ui/tooltip"
