+++ fdm-app/app/routes/farm.$b_id_farm.$calendar.balance.nitrogen._index.tsx
@@ -25,8 +25,8 @@ import { BufferStripInfo } from "~/components/blocks/balance/buffer-strip-info"
 import { FieldCultivationsBadge } from "~/components/blocks/balance/field-cultivations-badge"
 import { NitrogenBalanceChart } from "~/components/blocks/balance/nitrogen-chart"
 import { NitrogenBalanceFallback } from "~/components/blocks/balance/skeletons"
-import { CalculationRefreshBanner } from "~/components/blocks/calculation-refresh-banner"
-import { CalculationRefreshSpinner } from "~/components/blocks/calculation-refresh-spinner"
+import { CalculationRefreshBanner } from "~/components/blocks/calculation/calculation-refresh-banner"
+import { CalculationRefreshSpinner } from "~/components/blocks/calculation/calculation-refresh-spinner"
 import { Card, CardContent, CardDescription, CardHeader, CardTitle } from "~/components/ui/card"
 import { Tooltip, TooltipContent, TooltipTrigger } from "~/components/ui/tooltip"
 import { useAnalytics } from "~/hooks/use-analytics"
